x86/smp: Bring up secondary CPUs in two stages
Now that Thomas fixed the bringup we can do it in parallel for x86

We still can't do the callin part because it gets intertwined with
the TSC sync, so leave that part for later.

This reduces the time taken for bringup on my 28-thread Haswell
system by about a third.

Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>